добавил логи и проверку на null

+++ src/main/java/school/faang/user_service/service/RecommendationService.java
@@ -1,6 +1,7 @@
 package school.faang.user_service.service;
 
 import lombok.RequiredArgsConstructor;
+import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Service;
 import school.faang.user_service.dto.recommendation.RecommendationDto;
 import school.faang.user_service.entity.UserSkillGuarantee;
@@ -16,6 +17,7 @@ import school.faang.user_service.utils.ValidationRecommendationUtils;
 import java.util.ArrayList;
 import java.util.List;
 
+@Slf4j
 @RequiredArgsConstructor
 @Service
 public class RecommendationService {
@@ -40,7 +42,11 @@ public class RecommendationService {
     }
 
     public void delete(Long id) {
+        if (id == null) {
+            throw new DataValidationException("Id не должен быть null");
+        }
         recommendationRepository.deleteById(id);
+        log.info("Рекомендация id: {} была удалена", id);
     }
 
     private void deleteAllAndCreate(Recommendation recommendation) {
